A complete rewrite of the README
I am a dumb person... I used the Spanish translation of this README as the original basis for the English translation to save time, and that bit me in the butt. So, I have now retranslated EVERYTHING from scratch from the Japanese version using [deepl.com](deepl.com) as a suggestion from @jdegand. Thank you to him/her as well as @ThatOneCalculator for your help! I am dumb to say the least...

Co-Authored-By: J. Degand <70610011+jdegand@users.noreply.github.com>
Co-Authored-By: Kainoa Kanter <44733677+thatonecalculator@users.noreply.github.com>